refactor: 코드 정렬, 안쓰는 파일 삭제
